backend/agent/requirements.txt
livekit
livekit-agents[deepgram,openai,silero]
python-dotenv
aiohttp
